Hiring_microservices/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pydantic
pydantic[email]
python-multipart
jose
